Create Project5Part2
a C language program that inputs two ASCII hexadecimal digits followed by
the “+” symbol to indicate addition. Next, the user enters a second set of two hexadecimal digits
that are to be added to the first two hexadecimal digits. The addition result is shown on both the
hyperterminal display as well as the LCD display.